#599 Refactor runtime_layout into directory modules
